+++ source/mango/jpeg/jpeg_decode.cpp
@@ -11,7 +11,7 @@
 namespace mango {
 namespace jpeg {
 
-    static const int g_zigzag_table [] =
+    static const u8 g_zigzag_table [] =
     {
          0,  1,  8, 16,  9,  2,  3, 10,
         17, 24, 32, 25, 18, 11,  4,  5,
